+++ src/main/java/org/lenteam/colmen/repositories/SiteRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * @author Rinat
  */
 @Repository
-public interface SiteRepository extends CrudRepository<SiteEntity, Long> {
+public interface SiteRepository extends CrudRepository<SiteEntity, Integer> {
 }
